backend/requirements.txt
flask
flask-cors
google-generativeai
google-genai
openai
PyMuPDF
python-dotenv
requests
